combine blockMatcher and lazyBlockSig
